Remove tracetranslator.DetermineValueType, only used internally by Zipkin
