+++ translator/trace/zipkin/zipkinv2_to_traces.go
@@ -347,7 +347,7 @@ func tagsToAttributeMap(tags map[string]string, dest pdata.AttributeMap, parseSt
 		}
 
 		if parseStringTags {
-			switch tracetranslator.DetermineValueType(val) {
+			switch determineValueType(val) {
 			case pdata.AttributeValueINT:
 				iValue, _ := strconv.ParseInt(val, 10, 64)
 				dest.UpsertInt(key, iValue)
